# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Document Management System with AI and RFID support.",
        "title": "Kreatif DMS API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "romy@kreatif.id"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/batches": {
            "post": {
                "description": "Start a new batch for multiple uploads",
                "tags": [
                    "Batches"
                ],
                "summary": "Create Batch",
                "parameters": [
                    {
                        "description": "Total files in batch",
                        "name": "total_files",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_kreatif_dms-backend_internal_repository.ProcessingBatch"
                        }
                    }
                }
            }
        },
        "/batches/{id}": {
            "get": {
                "description": "Get progress and individual file status for a batch",
                "tags": [
                    "Batches"
                ],
                "summary": "Get Batch Status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Batch ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_kreatif_dms-backend_internal_service.BatchDetails"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_kreatif_dms-backend_internal_repository.Document": {
            "type": "object",
            "properties": {
                "batch_id": {
                    "type": "string"
                },
                "box_id": {
                    "type": "string"
                },
                "branch_id": {
                    "type": "string"
                },
                "checksum": {
                    "$ref": "#/definitions/pgtype.Text"
                },
                "circulation_id": {
                    "type": "string"
                },
                "company_id": {
                    "type": "string"
                },
                "created_at": {
                    "$ref": "#/definitions/pgtype.Timestamptz"
                },
                "current_version": {
                    "type": "integer"
                },
                "department_id": {
                    "type": "string"
                },
                "description": {
                    "$ref": "#/definitions/pgtype.Text"
                },
                "es_indexed": {
                    "$ref": "#/definitions/pgtype.Bool"
                },
                "extracted_text": {
                    "$ref": "#/definitions/pgtype.Text"
                },
                "file_name": {
                    "type": "string"
                },
                "file_path": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "is_ocr_processed": {
                    "$ref": "#/definitions/pgtype.Bool"
                },
                "metadata": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "mime_type": {
                    "type": "string"
                },
                "minio_bucket": {
                    "$ref": "#/definitions/pgtype.Text"
                },
                "ordner_id": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                },
                "rack_id": {
                    "type": "string"
                },
                "retention_expiry_date": {
                    "$ref": "#/definitions/pgtype.Date"
                },
                "retention_years": {
                    "$ref": "#/definitions/pgtype.Int4"
                },
                "sensitivity": {
                    "$ref": "#/definitions/pgtype.Text"
                },
                "status": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "$ref": "#/definitions/pgtype.Timestamptz"
                }
            }
        },
...
        "github_com_kreatif_dms-backend_internal_repository.ProcessingBatch": {
            "type": "object",
            "properties": {
                "created_at": {
                    "$ref": "#/definitions/pgtype.Timestamptz"
                },
                "department_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "intake_session_id": {
                    "type": "string"
                },
                "processed_files": {
                    "$ref": "#/definitions/pgtype.Int4"
                },
                "status": {
                    "$ref": "#/definitions/pgtype.Text"
                },
                "total_files": {
                    "type": "integer"
                },
                "updated_at": {
                    "$ref": "#/definitions/pgtype.Timestamptz"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "github_com_kreatif_dms-backend_internal_service.BatchDetails": {
            "type": "object",
            "properties": {
                "batch": {
                    "$ref": "#/definitions/github_com_kreatif_dms-backend_internal_repository.ProcessingBatch"
                },
                "documents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kreatif_dms-backend_internal_repository.Document"
                    }
                }
            }
        },
...
        "pgtype.Bool": {
            "type": "object",
            "properties": {
                "bool": {
                    "type": "boolean"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
        "pgtype.Date": {
            "type": "object",
            "properties": {
                "infinityModifier": {
                    "$ref": "#/definitions/pgtype.InfinityModifier"
                },
                "time": {
                    "type": "string"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
        "pgtype.InfinityModifier": {
            "type": "integer",
            "format": "int32",
            "enum": [
                1,
                0,
                -1
            ],
            "x-enum-varnames": [
                "Infinity",
                "Finite",
                "NegativeInfinity"
            ]
        },
        "pgtype.Int4": {
            "type": "object",
            "properties": {
                "int32": {
                    "type": "integer",
                    "format": "int32"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
        "pgtype.Text": {
            "type": "object",
            "properties": {
                "string": {
                    "type": "string"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
        "pgtype.Timestamptz": {
            "type": "object",
            "properties": {
                "infinityModifier": {
                    "$ref": "#/definitions/pgtype.InfinityModifier"
                },
                "time": {
                    "type": "string"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}